refactor: share Loom import processing
