+++ apps/web/__tests__/unit/loom-import.test.ts
@@ -468,6 +468,10 @@ describe("importFromLoom", () => {
 			videoId: "video-123",
 		});
 		expect(mockDb.delete).toHaveBeenCalledTimes(1);
+		expect(whereMock).toHaveBeenNthCalledWith(
+			2,
+			expect.arrayContaining([{ field: "id", value: "stale-row" }]),
+		);
 		expect(valuesMock).toHaveBeenCalledTimes(3);
 		expect(valuesMock).toHaveBeenNthCalledWith(
 			3,
@@ -480,6 +484,58 @@ describe("importFromLoom", () => {
 		);
 		expect(startMock).toHaveBeenCalledTimes(1);
 		expect(revalidatePathMock).toHaveBeenCalledWith("/dashboard/caps");
+	});
+
+	it("keeps the source claim transaction atomic on a duplicate collision", async () => {
+		whereMock.mockResolvedValueOnce([]);
+		valuesMock
+			.mockResolvedValueOnce(undefined)
+			.mockResolvedValueOnce(undefined)
+			.mockRejectedValueOnce(new Error("duplicate source claim"));
+
+		const fetchMock = vi.mocked(fetch);
+		fetchMock.mockImplementation(async (input) => {
+			const url = typeof input === "string" ? input : input.toString();
+
+			if (url.includes("/transcoded-url")) {
+				return {
+					ok: true,
+					status: 200,
+					text: async () =>
+						JSON.stringify({ url: "https://cdn.loom.com/video.mp4" }),
+				} as Response;
+			}
+
+			if (url === "https://www.loom.com/graphql") {
+				return {
+					ok: true,
+					json: async () => ({
+						data: { getVideo: { name: "Imported video" } },
+					}),
+				} as Response;
+			}
+
+			if (url.includes("/v1/oembed")) {
+				return {
+					ok: true,
+					json: async () => ({ duration: 42, width: 1920, height: 1080 }),
+				} as Response;
+			}
+
+			throw new Error(`Unexpected fetch: ${url}`);
+		});
+
+		const { importFromLoom } = await import("@/actions/loom");
+
+		await expect(
+			importFromLoom({
+				loomUrl: "https://www.loom.com/share/loom-abc1234567",
+				orgId: "org-1" as never,
+			}),
+		).rejects.toThrow("duplicate source claim");
+		expect(mockDb.transaction).toHaveBeenCalledTimes(1);
+		expect(valuesMock).toHaveBeenCalledTimes(3);
+		expect(startMock).not.toHaveBeenCalled();
 	});
 
 	it("rejects a CSV import when the current user is not an organization admin or owner", async () => {
@@ -904,7 +960,7 @@ describe("importFromLoom", () => {
 			],
 			error: undefined,
 		});
-		expect(mockDb.transaction).toHaveBeenCalledTimes(1);
+		expect(mockDb.transaction).toHaveBeenCalledTimes(2);
 		expect(valuesMock).toHaveBeenCalledWith(
 			expect.objectContaining({
 				name: "Sales Team",
